fix: no folder err handle
