+++ index.js
@@ -64,6 +64,12 @@ class FileManager {
     this.cd = commands.cd.bind(this)
     this.ls = commands.ls.bind(this)
     this.hash = commands.hash.bind(this)
+    this.cat = commands.cat.bind(this)
+    this.add = commands.add.bind(this)
+    this.rn = commands.rn.bind(this)
+    this.cp = commands.cp.bind(this)
+    this.mv = commands.mv.bind(this)
+    this.rm = commands.rm.bind(this)
   }
 
   startNewLine() {
@@ -114,6 +120,7 @@ class FileManager {
   }
 
   getAbsolutePath() {
+
     const folderInput = this.pathName
 
     if (folderInput) {
@@ -141,188 +148,27 @@ class FileManager {
   }
 
   cat() {
-    const path = this.getAbsolutePath()
-
-    this.stat(path, (err, stats) => {
-      if (err) {
-        this.err = `${path} is not a file`
-        return this.showError()
-      }
-      if (stats.isFile()) {
-        this.read(path)
-        .on('data', (chunk) => {
-          this.outStream(chunk)
-          this.startNewLine()
-          this.showDirectory()
-        })
-      }
-    })
+    this.cat()
   }
 
   add() {
-    const path = this.getCurrentDirName() + '/' + this.pathName
-
-    this.writer(path)
-      this.outStream(`${this.pathName} has just created`)
-      this.startNewLine()
-      this.showDirectory()
+    this.add()
   }
 
   rn() {
-    const path = this.getAbsolutePath()
-
-    this.stat(path, (err, stats) => {
-      if (err) {
-        this.err = err
-        return this.showError()
-      }
-
-      if (!stats.isFile()) {
-        this.err = `${path} is not a file`
-        return this.showError()
-      }
-
-      const directory = this.parsePath(path).dir
-
-      const newFilePath = this.resolvePath(directory, this.newFileName)
-      const writeable = this.writer(newFilePath)
-      this.read(path)
-      .on('data', (chunk) => {
-        writeable.write(chunk)
-      })
-      .on('close', () => {
-        this.unlink(path, (err) => {
-          if (err) {
-            this.err = err
-            return this.showError()
-          }
-        })
-        this.outStream(`${this.pathName} was just renamed \n the changes will came into power after end of the session`)
-        this.startNewLine()
-        this.showDirectory()
-      })
-    })
+    this.rn()
   }
 
   cp() {
-    const path = this.getAbsolutePath()
-
-    this.stat(path, (err, stats) => {
-      if (err) {
-        this.err = err
-        return this.showError()
-      }
-
-      if (!stats.isFile()) {
-        this.err = `${path} is not a file`
-        return this.showError()
-      }
-
-
-      const copyFileName = this.parsePath(path).base
-      const newFilePath = this.resolvePath(this.newFileDirectory, copyFileName)
-
-      this.stat(newFilePath, (err, stats) => {
-        if (err && !err.code === 'ENOENT') {
-          this.err = err
-          return this.showError()
-        }
-
-        if (stats !== undefined && stats.isFile()) {
-          this.err = `${path} already exists`
-          return this.showError()
-        }
-      })
-
-      const writeable = this.writer(newFilePath)
-      this.read(path)
-      .on('data', (chunk) => {
-        writeable.write(chunk)
-      })
-      .on('close', () => {
-        this.outStream(`${this.pathName} was just copied`)
-        this.startNewLine()
-        this.showDirectory()
-      })
-    })
+    this.cp()
   }
 
   mv() {
-    const path = this.getAbsolutePath()
-
-    this.stat(path, (err, stats) => {
-      if (err) {
-        this.err = err
-        return this.showError()
-      }
-
-      if (!stats.isFile()) {
-        this.err = `${path} is not a file`
-        return this.showError()
-      }
-
-
-      const copyFileName = this.parsePath(path).base
-      const newFilePath = this.resolvePath(this.newFileDirectory, copyFileName)
-
-      this.stat(newFilePath, (err, stats) => {
-        if (err && !err.code === 'ENOENT') {
-          this.err = err
-          return this.showError()
-        }
-
-        if (stats !== undefined && stats.isFile()) {
-          this.err = `${path} already exists`
-          return this.showError()
-        }
-      })
-
-      const writeable = this.writer(newFilePath)
-      this.read(path)
-      .on('data', (chunk) => {
-        writeable.write(chunk)
-      })
-      .on('close', () => {
-        this.unlink(path, (err) => {
-          if (err) {
-            this.err = err
-            return this.showError()
-          }
-        })
-        this.outStream(`
-        ${this.pathName} was just moved into ${this.newFileDirectory} 
-        the changes will came into power after end of the session
-        `)
-        this.startNewLine()
-        this.showDirectory()
-      })
-    })
+    this.mv()
   }
 
   rm() {
-    const path = this.getAbsolutePath()
-
-    this.stat(path, (err, stats) => {
-      if (err) {
-        this.err = err
-        return this.showError()
-      }
-
-      if (!stats.isFile()) {
-        this.err = `${path} is not a file`
-        return this.showError()
-      }
-
-      this.unlink(path, (err) => {
-        if (err) {
-          this.err = err
-          return this.showError()
-        }
-        })
-        this.outStream(`${this.pathName} was just deleted \n the changes will came into power after end of the session`)
-        this.startNewLine()
-        this.showDirectory()
-    })
+    this.rm()
   }
 
   logEol() {
@@ -392,6 +238,12 @@ class FileManager {
       const pathNames = userInput.replace(/^rn /, '').split(' ')
       const currentPath = pathNames[0]
       const newFileName = pathNames[1]
+
+      if (!newFileName) {
+        this.err = `incorrect destination folder`
+        return this.showError()
+      }
+
       this.pathName = currentPath
       this.newFileName = newFileName
     }
@@ -402,6 +254,12 @@ class FileManager {
       const pathNames = userInput.replace(/^cp /, '').split(' ')
       const currentPath = pathNames[0]
       const newFileName = pathNames[1]
+ 
+      if (!newFileName) {
+        this.err = `incorrect destination folder`
+        return this.showError()
+      }
+
       this.pathName = currentPath
       this.newFileDirectory = newFileName
     }
@@ -412,6 +270,12 @@ class FileManager {
       const pathNames = userInput.replace(/^mv /, '').split(' ')
       const currentPath = pathNames[0]
       const newFileName = pathNames[1]
+
+      if (!newFileName) {
+        this.err = `incorrect destination folder`
+        return this.showError()
+      }
+
       this.pathName = currentPath
       this.newFileDirectory = newFileName
     }
@@ -420,6 +284,11 @@ class FileManager {
       const userInput = this.input
       this.input = 'rm'
       this.pathName = userInput.replace(/^rm /, '')
+
+      if (!pathName) {
+        this.err = `incorrect filename`
+        return this.showError()
+      }
     }
 
     if (/^os /i.test(this.input)) {
@@ -515,7 +384,7 @@ class FileManager {
             case 'rm':
               this.rm()
               break
-            case '--eol':
+            case '--EOL':
               this.logEol()
               break
             case '--cpus':
